requirements.txt
anyio==3.7.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
dnspython==2.3.0
exceptiongroup==1.1.2
fastapi==0.99.1
google-api-core==2.11.1
google-api-python-client==2.93.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
gspread==5.10.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
numpy==1.25.0
oauth2client==4.1.3
oauthlib==3.2.2
pandas==2.0.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.11
PyJWT==2.7.0
pymongo==4.4.0
pyOpenSSL==23.2.0
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.22.0
azure-core==1.26.4
azure-cosmos==4.3.1
# certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
distlib==0.3.6
filelock==3.12.0
idna==3.4
pip-upgrade==0.0.6
platformdirs==3.2.0
pur==7.1.0
# requests==2.28.2
# six==1.16.0
# typing_extensions==4.5.0
# urllib3==1.26.15
virtualenv==20.22.0
virtualenv-clone==0.5.7
